[Feature #111504652] Restructure File/ Folder, Cleanup Code, Create api/v1/auth/ routes
