+++ codango/codango/settings/base.py
@@ -81,7 +81,8 @@ INSTALLED_APPS = (
     'djangobower',
     'endless_pagination',
     'djcelery',
-    'api'
+    'rest_framework',
+    'api',
 )
 
 MIDDLEWARE_CLASSES = (
