+++ codango/codango/urls.py
@@ -20,15 +20,15 @@ import userprofile.urls
 import resources.urls
 import comments.urls
 import pairprogram.urls
-import api.urls
+import api.urls.urls
 
 
 urlpatterns = [
     url(r'^', include(account.urls)),
+    url(r'^api/v1/', include(api.urls.urls)),
     url(r'^resource/', include(resources.urls)),
     url(r'^user/', include(userprofile.urls)),
     url(r'^comment/', include(comments.urls)),
-    url(r'^api/v1/', include(api.urls)),
     url(r'^pair/', include(pairprogram.urls)),
     url(r'^admin/', include(admin.site.urls)),
 ]
